# feat: Complete Phase 1.1 – Error Handling Validation, Performance Benchmarking, and Documentation Updates