--- a/src/app.controller.ts
+++ b/src/app.controller.ts
@@ -8,7 +8,9 @@ export class AppController {
   @Get()
   getHello(): string {
     // 🚀 HOT RELOAD WORKING - Manual rebuild at 12:58 PM
+    // 🔥 Testing hot reload at 1:00 PM after network policy fix
+    // 🧪 VALIDATION TEST: Hot reload check at 1:01 PM
     console.log('🎯 Hello endpoint called at:', new Date().toISOString());
-    return this.appService.getHello();
+    return this.appService.getHello() + ' - Hot reload test v3';
   }
 }