… .dev-cache/build-cache.json
… .env.template
… docker-compose.dev.yml
… docs/DEVELOPMENT.md
… docs/PERFORMANCE_VALIDATION.md
… docs/PLANNING.md
… k8s/base/configmap.yaml
… k8s/base/deployment.yaml
… k8s/base/kafka.yaml
… k8s/base/kustomization.yaml
… k8s/base/migration-job.yaml
… k8s/overlays/dev-minimal/configmap-patch.yaml
… k8s/overlays/dev-minimal/deployment-patch.yaml
… k8s/overlays/dev-minimal/deployment.yaml
… k8s/overlays/dev-minimal/kustomization.yaml
… k8s/overlays/dev-minimal/namespace.yaml
… k8s/overlays/dev-minimal/networkpolicy-patch.yaml
… k8s/overlays/dev-minimal/postgres.yaml
… k8s/overlays/dev-minimal/service.yaml
… k8s/overlays/dev/configmap-patch.yaml
… k8s/overlays/dev/kustomization.yaml
… package.json
… scripts/performance-benchmark.js
… skaffold.yaml
… src/app.controller.spec.ts
… src/common/decorators/api-key-auth.decorator.spec.ts
… src/common/decorators/request-id.decorator.spec.ts
… src/common/exceptions/wal.exceptions.spec.ts
… src/common/filters/all-exceptions.filter.spec.ts
… src/common/interceptors/logging.interceptor.spec.ts